Complete substring project
Achieved this by first checking if the string is
a single word or a sentence. If true, then split
the string into an array of characters and pass
it as arguments for hashing function. If false,
i split the sentence to an array of strings then
with #each loop i get single words and split them
into an array of characters and pass them into
hashing function.

hashing func: takes in a character array,
dictionary and a result hash. Looping through
the dictionary i change each word into character
array and check if all the characters are present
in the user given character array. If true, add
the word into the hash with the count. If false,
continue.